requirements.txt
# 核心依赖
pm4py==2.7.11
python-dotenv==1.0.0
requests==2.31.0
flask==2.3.3
numpy==1.26.0
pandas==2.1.1

# 机器学习依赖
torch>=2.4.1  # 只指定主版本，不指定CUDA后缀
scikit-learn==1.3.0
transformers==4.34.0
tensorboard==2.14.1

# 联邦学习依赖
flwr==1.5.0
tenseal==0.3.16  

# 开发依赖
pytest==7.4.3
black==23.11.0
